ai/pydantic-ai/rag-func/pyproject.toml
[project]
name = "rag-func"
version = "0.1.0"
description = "Basic AI with primitive RAQ"
readme = "../../README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.0.18",
    "python-dotenv>=1.0.1",
]
